Change type of Rating property to decimal in Review entity

+++ BidUp.BusinessLogic/DTOs/ReviewsDTOs/ReviewResponse.cs
@@ -3,7 +3,7 @@ namespace BidUp.BusinessLogic.DTOs.ReviewsDTOs;
 public class ReviewResponse
 {
     public int Id { get; init; }
-    public int Rating { get; init; }
+    public decimal Rating { get; init; }
     public string? Comment { get; init; }
     public DateTime CreatedAt { get; init; }
     public DateTime UpdatedAt { get; init; }
